Type updates. R

+++ arduino/Different motors/arduino_R.ino
@@ -29,10 +29,10 @@ String incomingCommand = "";
 //0: Stand still
 //1: Move to targetLocationNumber
 //2: Reset
-int missionIndex = 0;
+int16_t missionIndex = 0;
 
 //used for resume
-int savedMissionIndex = 0;
+int16_t savedMissionIndex = 0;
 
 
 
@@ -42,16 +42,16 @@ int savedMissionIndex = 0;
 float delayTime = 20.0;
 
 
-volatile long locationNumber = 0; //Goes from 0 to somewhere between 10k and 20k
-long prevLocationNumber = locationNumber;
-long longagoPositionOne = locationNumber; //0 to 1.5s ago
-long longagoPositionTwo = locationNumber; //1.5 to 3s ago
-long longagoPositionThree = locationNumber; //3 to 4.5s ago
-int loopsPerLongagoPositionUpdate = 1500/delayTime;
-int counter = 0;
+volatile int32_t locationNumber = 0; //Goes from 0 to somewhere between 10k and 20k
+int32_t prevLocationNumber = locationNumber;
+int32_t longagoPositionOne = locationNumber; //0 to 1.5s ago
+int32_t longagoPositionTwo = locationNumber; //1.5 to 3s ago
+int32_t longagoPositionThree = locationNumber; //3 to 4.5s ago
+int16_t loopsPerLongagoPositionUpdate = 1500/delayTime;
+int16_t counter = 0;
 
 
-long targetLocationNumber = 0;
+int32_t targetLocationNumber = 0;
 
 
 
@@ -66,8 +66,8 @@ float P = 0.00035;
 float I = 0.000000005;
 float D = 0.07;
 float generalSpeedFactor = 0.35; //0.45
-int forwardsMargin = 25;
-int backwardsMargin = 5;
+int16_t forwardsMargin = 25;
+int16_t backwardsMargin = 5;
 
 
 
@@ -75,14 +75,14 @@ float antistuckCurrentPWMBonus = 0;
 float integral = 0;
 
 float PWMFraction = 0.0;
-int movementDir = 0; //-1 for backwards, +1 for forwards, 0 for standing still
+int16_t movementDir = 0; //-1 for backwards, +1 for forwards, 0 for standing still
 
 
 
 
 //used for detecting loose cables
-int locationNumberPreviousVarvInterrupt = locationNumber;
-int numberOfSusVarvInterrupts = 0;
+int16_t locationNumberPreviousVarvInterrupt = locationNumber;
+int16_t numberOfSusVarvInterrupts = 0;
 
 
 
@@ -150,7 +150,7 @@ void loop() {
       } else {
         missionIndex = 1;
 		savedMissionIndex = 1;
-        targetLocationNumber = atoi(incomingCommand.c_str());
+        targetLocationNumber = atol(incomingCommand.c_str());
       }
